requirements.txt
fastapi==0.110.0
fastapi-utils==0.2.1

langchain~=0.1.13
pydantic~=1.10.15
openai~=1.30.4
pytube~=15.0.0
whisper~=1.1.10
PyYAML~=6.0.1
uvicorn~=0.30.0